refactor layout structure and components
